requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dgl==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
littleutils==0.2.4
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.24.3
ogb==1.3.6
openai==0.28.1
outdated==0.2.2
packaging==24.2
pandas==2.0.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==5.8.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rdkit==2022.3.5
requests==2.32.3
scikit-learn==1.2.2
scipy==1.13.1
setuptools==75.1.0
six==1.17.0
stack_data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.1.0
tornado==6.4.2
tqdm==4.65.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.44.0
yarl==1.18.3
zipp==3.21.0
